Identify version from ChangeLog & Documentation.html
PHPMyAdmin stores its changes in a file named ChangeLog. The script
first checks in this file to identify the version.
If the file has been removed from the server, it checks in the file
Documentation.html. This file is present in old versions of PHPMyAdmin
but it's still possible to find these versions online.